+++ source/iterative_methods.h
@@ -1,16 +1,16 @@
 #include "matrixes.h"
 
-std::vector<float> simple_iteration(const CSRMatrix &A,
-                                    const std::vector<float> &b,
-                                    const std::vector<float> &x0, const float t,
-                                    const float tolerance);
+std::vector<double> simple_iteration(const CSRMatrix &A,
+                                     const std::vector<double> &b,
+                                     const std::vector<double> &x0,
+                                     const double t, const double tolerance);
 
-std::vector<float> jacobi_iteration(const CSRMatrix &A,
-                                    const std::vector<float> &b,
-                                    const std::vector<float> &x0,
-                                    const float tolerance);
+std::vector<double> jacobi_iteration(const CSRMatrix &A,
+                                     const std::vector<double> &b,
+                                     const std::vector<double> &x0,
+                                     const double tolerance);
 
-std::vector<float> gauss_seidel_iteration(const CSRMatrix &A,
-                                          const std::vector<float> &b,
-                                          const std::vector<float> &x0,
-                                          const float tolerance);
+std::vector<double> gauss_seidel_iteration(const CSRMatrix &A,
+                                           const std::vector<double> &b,
+                                           const std::vector<double> &x0,
+                                           const double tolerance);
