+++ source/solvers.cpp
@@ -1,10 +1,10 @@
 #include "matrixes.h"
 #include <vector>
 
-std::vector<float> solve(const TridiagonalMatrix &matrix,
-                         const std::vector<float> &column) {
+std::vector<double> solve(const TridiagonalMatrix &matrix,
+                          const std::vector<double> &column) {
     int n = matrix.get_size() - 1;
-    std::vector<float> p_vector{0}, q_vector{0};
+    std::vector<double> p_vector{0}, q_vector{0};
     p_vector.reserve(matrix.get_size());
     q_vector.reserve(matrix.get_size());
     for (int i = 0; i < n + 1; ++i) {
@@ -13,7 +13,7 @@ std::vector<float> solve(const TridiagonalMatrix &matrix,
         q_vector.push_back((column[i] - matrix(i, i - 1) * q_vector[i]) /
                            (matrix(i, i - 1) * p_vector[i] + matrix(i, i)));
     }
-    std::vector<float> solution(n + 1);
+    std::vector<double> solution(n + 1);
     solution[n] = (column[n] - matrix(n, n - 1) * q_vector[n]) /
                   (matrix(n, n - 1) * p_vector[n] + matrix(n, n));
     for (int i = n - 1; i >= 0; --i) {
