+++ tests/dense_test.cpp
@@ -10,11 +10,11 @@ TEST(DenseTest, CheckItemGetting) {
 }
 
 TEST(DenseTest, CheckMatrixVectorMultiplication) {
-    const std::vector<float> column1 = {1, -3, 2};
-    const std::vector<float> res1 = {9, 3, 10};
+    const std::vector<double> column1 = {1, -3, 2};
+    const std::vector<double> res1 = {9, 3, 10};
     EXPECT_EQ(slae1 * column1, res1);
-    const std::vector<float> column2 = {0, 0, 0};
-    const std::vector<float> res2 = {0, 0, 0};
+    const std::vector<double> column2 = {0, 0, 0};
+    const std::vector<double> res2 = {0, 0, 0};
     EXPECT_EQ(slae1 * column2, res2);
 }
 
@@ -25,12 +25,12 @@ TEST(DenseTest, CheckRowNumber) {
 }
 
 TEST(DenseTest, CheckColumnGetting) {
-    const std::vector<float> col1 = {1, 0, 0};
+    const std::vector<double> col1 = {1, 0, 0};
     EXPECT_EQ(slae1.get_column(0), col1);
 }
 
 TEST(DenseTest, CheckRowGetting) {
-    const std::vector<float> row1 = {0, 3, 6};
+    const std::vector<double> row1 = {0, 3, 6};
     EXPECT_EQ(slae1.get_row(1), row1);
 }
