+++ tests/tridiagonal_test.cpp
@@ -11,16 +11,16 @@ TEST(TridiagonalTest, CheckItemGetting) {
 }
 
 TEST(TridiagonalTest, CheckSolving) {
-    const float ERROR = 0.001;
+    const double ERROR = 0.001;
 
-    std::vector<float> res1 = solve(slae1, {0.04, 0.04, 0.04, 0.04});
-    const std::vector<float> solution1{-0.08, -0.12, -0.12, -0.08};
+    std::vector<double> res1 = solve(slae1, {0.04, 0.04, 0.04, 0.04});
+    const std::vector<double> solution1{-0.08, -0.12, -0.12, -0.08};
     for (int i = 0; i < res1.size(); ++i) {
         EXPECT_NEAR(res1[i], solution1[i], ERROR);
     }
 
-    std::vector<float> res2 = solve(slae2, {21, 69, 34, 22});
-    const std::vector<float> solution2{3, 5, 4, 2};
+    std::vector<double> res2 = solve(slae2, {21, 69, 34, 22});
+    const std::vector<double> solution2{3, 5, 4, 2};
     for (int i = 0; i < res2.size(); ++i) {
         EXPECT_NEAR(res2[i], solution2[i], ERROR);
     }
